refactor(tooltiptable): rename component and story, remove table from Tooltip stories
